fix:Add version code and del warning

+++ sysapp/monitor/vuicc/src/vuicc_apdu.c
@@ -80,7 +80,7 @@ int32_t vuicc_lpa_cmd(const uint8_t *data, uint16_t data_len, uint8_t *rsp, uint
 static int32_t vuicc_get_atr(uint8_t *atr, uint8_t *atr_size)
 {
     rt_mutex_lock(&g_apdu_mutex);
-    cos_client_reset((uint8_t *)atr, (uint16_t *)atr_size);
+    cos_client_reset((uint8_t *)atr, (uint8_t *)atr_size);
     rt_mutex_unlock(&g_apdu_mutex);
 }
